requirements.txt
streamlit>=1.30
pytest>=7.0
anthropic>=0.7.0
google-genai>=1.0.0
openai>=1.0.0
sentence-transformers>=2.2.0
transformers>=4.41.0,<5.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
